Moving fake heimdal service client etc. to /test
and completing ExtErrorReportControllerTest
