+++ kifi-backend/modules/common/test/com/keepit/heimdal/FakeHeimdalServiceClientImpl.scala
@@ -1,0 +1,35 @@
+package com.keepit.heimdal
+
+import com.keepit.model.User
+import com.keepit.common.db.Id
+import com.keepit.common.service.{ServiceClient, ServiceType}
+import com.keepit.common.logging.Logging
+import com.keepit.common.routes.Heimdal
+import com.keepit.common.healthcheck.AirbrakeNotifier
+import com.keepit.common.net.HttpClient
+import com.keepit.common.zookeeper.ServiceCluster
+
+import scala.concurrent.{Future, Promise}
+
+import play.api.libs.json.{JsArray, Json, JsObject}
+import play.api.libs.concurrent.Execution.Implicits.defaultContext
+
+import com.google.inject.Inject
+
+
+class FakeHeimdalServiceClientImpl(val airbrakeNotifier: AirbrakeNotifier) extends HeimdalServiceClient{
+  val serviceCluster: ServiceCluster = new ServiceCluster(ServiceType.TEST_MODE)
+  protected def httpClient: com.keepit.common.net.HttpClient = ???
+
+  var eventsRecorded : Int = 0
+
+  def trackEvent(event: UserEvent): Unit = synchronized{
+    eventsRecorded =  eventsRecorded + 1
+  }
+
+  def eventCount: Int = eventsRecorded
+
+  def getMetricData(name: String): Future[JsObject] = Promise.successful(Json.obj()).future
+
+  def updateMetrics(): Unit = {}
+}
